:sparkles: Create credentials endpoints namespace
